Move to CPM and adopt canonical project structure
