+++ .cmake-format.yaml
@@ -1,73 +1,22 @@
 parse:
   additional_commands:
-    foo:
-      flags:
-        - BAR
-        - BAZ
+    add_target:
       kwargs:
-        HEADERS: "*"
-        SOURCES: "*"
-        DEPENDS: "*"
-  override_spec: {}
-  vartags: []
-  proptags: []
+        NAME: 1
+        TYPE: 1
+        SOURCES:
+          flags: [PUBLIC, PRIVATE]
+        INCLUDE_DIRS: "*"
+        LINK_LIBS: "*"
+    target_set_warnings:
+      kwargs:
+        TARGET: "*"
+        ENABLE: "*"
+        AS_ERRORS: "*"
 format:
-  disable: false
-  line_width: 80
+  line_width: 120
   tab_size: 4
-  use_tabchars: false
-  fractional_tab_policy: use-space
-  max_subgroups_hwrap: 2
-  max_pargs_hwrap: 6
-  max_rows_cmdline: 2
-  separate_ctrl_name_with_space: false
-  separate_fn_name_with_space: false
-  dangle_parens: false
-  dangle_align: prefix
-  min_prefix_chars: 4
-  max_prefix_chars: 10
-  max_lines_hwrap: 2
-  line_ending: unix
-  command_case: canonical
-  keyword_case: unchanged
-  always_wrap: []
-  enable_sort: true
-  autosort: false
-  require_valid_layout: false
-  layout_passes: {}
+  max_pargs_hwrap: 4
 markup:
-  bullet_char: "*"
-  enum_char: .
-  first_comment_is_literal: false
-  literal_comment_pattern: null
-  fence_pattern: ^\s*([`~]{3}[`~]*)(.*)$
-  ruler_pattern: ^\s*[^\w\s]{3}.*[^\w\s]{3}$
-  explicit_trailing_pattern: "#<"
-  hashruler_min_length: 10
-  canonicalize_hashrulers: true
-  enable_markup: true
-lint:
-  disabled_codes: []
-  function_pattern: "[0-9a-z_]+"
-  macro_pattern: "[0-9A-Z_]+"
-  global_var_pattern: "[A-Z][0-9A-Z_]+"
-  internal_var_pattern: _[A-Z][0-9A-Z_]+
-  local_var_pattern: "[a-z][a-z0-9_]+"
-  private_var_pattern: _[0-9a-z_]+
-  public_var_pattern: "[A-Z][0-9A-Z_]+"
-  argument_var_pattern: "[a-z][a-z0-9_]+"
-  keyword_pattern: "[A-Z][0-9A-Z_]+"
-  max_conditionals_custom_parser: 2
-  min_statement_spacing: 1
-  max_statement_spacing: 2
-  max_returns: 6
-  max_branches: 12
-  max_arguments: 5
-  max_localvars: 15
-  max_statements: 50
-encode:
-  emit_byteorder_mark: false
-  input_encoding: utf-8
-  output_encoding: utf-8
-misc:
-  per_command: {}
+  first_comment_is_literal: true
+  enable_markup: false
