+++ .yamllint.yaml
@@ -1,6 +1,7 @@
----
 extends: default
 
 rules:
+  document-start:
+    present: false
   line-length:
     max: 100
