+++ ball_sort/ball_sort/puzzle/puzzle_make_tubes.cpp
@@ -1,4 +1,4 @@
-#include "ball_sort/exceptions/illegal_puzzle_exception.hpp"
+#include "ball_sort/illegal_puzzle_exception.hpp"
 #include "ball_sort/puzzle.hpp"
 #include <fmt/core.h>
 #include <sstream>
