+++ ball_sort/ball_sort/tube.cpp
@@ -1,6 +1,6 @@
 #include "ball_sort/tube.hpp"
-#include <ball_sort/exceptions/illegal_move_exception.hpp>
-#include <ball_sort/exceptions/illegal_puzzle_exception.hpp>
+#include <ball_sort/illegal_move_exception.hpp>
+#include <ball_sort/illegal_puzzle_exception.hpp>
 #include <fmt/core.h>
 
 namespace ballsort {
