+++ ball_sort/tests/puzzle/test_puzzle_do_move.cpp
@@ -1,4 +1,4 @@
-#include "ball_sort/exceptions/illegal_move_exception.hpp"
+#include "ball_sort/illegal_move_exception.hpp"
 #include "ball_sort/puzzle.hpp"
 #include <gtest/gtest.h>
