+++ cmake/Sanitisers.cmake
@@ -6,21 +6,17 @@ function(target_add_address_sanitiser TARGET)
         return()
     endif()
 
-    if(CMAKE_CXX_COMPILER_ID MATCHES "Clang" OR CMAKE_CXX_COMPILER_ID MATCHES
-                                                "GNU")
+    if(CMAKE_CXX_COMPILER_ID MATCHES "Clang" OR CMAKE_CXX_COMPILER_ID MATCHES "GNU")
         target_compile_options(${TARGET} PRIVATE "-fno-omit-frame-pointer")
         target_link_options(${TARGET} PRIVATE "-fno-omit-frame-pointer")
 
-        message(
-            STATUS "[SANITISERS] Address sanitiser active for target: ${TARGET}"
-        )
         target_compile_options(${TARGET} PRIVATE "-fsanitize=address")
         target_link_options(${TARGET} PRIVATE "-fsanitize=address")
 
+        message(STATUS "[${TARGET}] Address sanitiser active")
+
     elseif(CMAKE_CXX_COMPILER_ID MATCHES "MSVC")
-        message(
-            STATUS "[SANITISERS] Address sanitiser active for target: ${TARGET}"
-        )
+        message(STATUS "[${TARGET}] Address sanitiser active")
         target_add_compile_options(${TARGET} PRIVATE "/fsanitize=address")
 
     else()
@@ -38,21 +34,16 @@ function(target_add_undefined_behaviour_sanitiser TARGET)
         return()
     endif()
 
-    if(CMAKE_CXX_COMPILER_ID MATCHES "Clang" OR CMAKE_CXX_COMPILER_ID MATCHES
-                                                "GNU")
+    if(CMAKE_CXX_COMPILER_ID MATCHES "Clang" OR CMAKE_CXX_COMPILER_ID MATCHES "GNU")
         target_compile_options(${TARGET} PRIVATE "-fno-omit-frame-pointer")
         target_link_options(${TARGET} PRIVATE "-fno-omit-frame-pointer")
 
-        message(
-            STATUS
-                "[SANITISERS] Undefined behaviour sanitiser active for target: ${TARGET}"
-        )
+        message(STATUS "[${TARGET}] Undefined behaviour sanitiser active")
         target_compile_options(${TARGET} PRIVATE "-fsanitize=undefined")
         target_link_options(${TARGET} PRIVATE "-fsanitize=undefined")
 
     elseif(CMAKE_CXX_COMPILER_ID MATCHES "MSVC")
-        message(
-            WARNING "Undefined behaviour sanitiser not implemented for MSVC")
+        message(WARNING "Undefined behaviour sanitiser not implemented for MSVC")
 
     else()
         message(WARNING "Compiler not supported for sanitisers")
@@ -68,24 +59,19 @@ function(target_add_thread_sanitiser TARGET)
         return()
     endif()
 
-    if(CMAKE_CXX_COMPILER_ID MATCHES "Clang" OR CMAKE_CXX_COMPILER_ID MATCHES
-                                                "GNU")
+    if(CMAKE_CXX_COMPILER_ID MATCHES "Clang" OR CMAKE_CXX_COMPILER_ID MATCHES "GNU")
         target_compile_options(${TARGET} PRIVATE "-fno-omit-frame-pointer")
         target_link_options(${TARGET} PRIVATE "-fno-omit-frame-pointer")
 
         if(ENABLE_SANITISE_ADDRESS)
-            message(
-                WARNING "Thread sanitiser does not work with address sanitiser")
+            message(WARNING "Thread sanitiser does not work with address sanitiser")
         else()
-            message(
-                STATUS
-                    "[SANITISERS] Thread sanitiser active for target: ${TARGET}"
-            )
+            message(STATUS "[${TARGET}] Thread sanitiser active")
             add_compile_options("-fsanitize=thread")
             add_link_options("-fsanitize=thread")
         endif()
 
-    elseif(CMAKE_CXX_COMPILER_ID MATCHES PRIVATE "MSVC")
+    elseif(CMAKE_CXX_COMPILER_ID MATCHES "MSVC")
         message(WARNING "Thread sanitiser not implemented for MSVC")
 
     else()
