+++ scripts/release.sh
@@ -3,6 +3,5 @@
 set -e
 
 rm -rf build
-conan install . -s build_type=Release --build=missing -s compiler.cppstd=20
-cmake -DENABLE_WARNINGS_AS_ERRORS=OFF --preset conan-release
-cmake --build --preset conan-release
+cmake -B build -DCMAKE_BUILD_TYPE=Release -DCMAKE_CXX_COMPILER_LAUNCHER=ccache
+cmake --build build
